requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-extensions==3.2.0
django-oauth-toolkit==2.1.0
djangorestframework==3.13.1
Faker==14.1.0
gunicorn==20.1.0
idna==3.3
jwcrypto==1.3.1
mysql-client==0.0.1
mysqlclient==2.1.1
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
wrapt==1.14.1
